+++ nova-app/src/features/trend/api/trend.ts
@@ -1,7 +1,6 @@
 import { axiosInstance } from '@/shared/api';
 import { ApiSuccessResponse } from '@/shared/types';
 import { KeywordTopDTO, skilltopDTO } from '@/features/trend/types/api';
-import { useSuspenseQuery } from '@tanstack/react-query';
 export const getKeywordTop = async (): Promise<ApiSuccessResponse<KeywordTopDTO>> => {
   const { data } = await axiosInstance.get<ApiSuccessResponse<KeywordTopDTO>>(
     `/api/trends/keywords/keywordtop`,
@@ -9,25 +8,9 @@ export const getKeywordTop = async (): Promise<ApiSuccessResponse<KeywordTopDTO>
   return data;
 };
 
-export const useGetKeywordTop = () => {
-  return useSuspenseQuery({
-    queryKey: ['keywords'],
-    queryFn: getKeywordTop,
-    select: (res) => res.data,
-  });
-};
-
 export const getInterestSkillTop = async (): Promise<ApiSuccessResponse<skilltopDTO>> => {
   const { data } = await axiosInstance.get<ApiSuccessResponse<skilltopDTO>>(
     `/api/trends/interests/skilltop`,
   );
   return data;
-};
-
-export const useGetInterestSkillTop = () => {
-  return useSuspenseQuery({
-    queryKey: ['interests'],
-    queryFn: getInterestSkillTop,
-    select: (res) => res.data,
-  });
 };
