+++ nova-app/src/features/trend/hooks/useGetInterestSkillTop.ts
@@ -1,0 +1,10 @@
+import { getInterestSkillTop } from '@/features/trend/api/trend';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+export const useGetInterestSkillTop = () => {
+  return useSuspenseQuery({
+    queryKey: ['interests'],
+    queryFn: getInterestSkillTop,
+    select: (res) => res.data,
+  });
+};
