+++ nova-app/src/features/trend/hooks/useGetKeywordTop.ts
@@ -1,0 +1,10 @@
+import { getKeywordTop } from '@/features/trend/api/trend';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+export const useGetKeywordTop = () => {
+  return useSuspenseQuery({
+    queryKey: ['keywords'],
+    queryFn: getKeywordTop,
+    select: (res) => res.data,
+  });
+};
